Refactor numba.rewrites as a package
